# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API StageSeller",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Api StageSeller"
        }
    ],
    "paths": {
...
        "/api/matches": {
            "post": {
                "tags": [
                    "Match"
                ],
                "operationId": "b1ee11f72732c34254f248bac13f5b29",
                "parameters": [
                    {
                        "name": "receveir",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created / Returns a like for the user who receives it",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Match-add"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid JWT || Expired JWT Token code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "This value should not be blank"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/matches/{id}": {
            "put": {
                "tags": [
                    "Match"
                ],
                "operationId": "768e78b30a33107c2695b00fc116a84b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Match-edit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Match not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid JWT || Expired JWT Token code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Only the receiver can update the match code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "receiver: This value should not be blank"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Match"
                ],
                "operationId": "423d4e8d95d241d45828e531e2beeff0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content / Delete of a match user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Match-delete"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Match not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid JWT || Expired JWT Token code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Match-add": {
                "description": "Match-add",
                "properties": {
                    "receiver": {
                        "type": "integer"
                    },
                    "isMutual": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Match-edit": {
                "description": "Match-edit",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Match-delete": {
                "description": "Match-delete",
                "properties": {
                    "match": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}